requirements.txt
fastapi
aioredis
python-dotenv
pytest
aiohttp
uvicorn
httpx
pytest-asyncio